+++ src/lib/components/Currency/Currency.svelte
@@ -1,12 +1,16 @@
 <script type="ts">
 	export let amount: number;
-	export let currency: 'FLOW' | 'FUSD' | 'USD';
+	export let currency: 'FLOW' | 'FUSD' | 'USD' | string;
 	export let fontSize: string = '1rem';
 	export let color: 'heading' | 'text' = 'text';
 </script>
 
 <div style={`font-size: ${fontSize}; color: var(--clr-${color}-main)`}>
-	<span class="amount">{`$${amount.toLocaleString()}`}</span>
+	{#if amount >= 0}
+		<span class="amount">{`$${amount.toLocaleString()}`}</span>
+	{:else}
+		<span class="amount">{`-$${Math.abs(amount).toLocaleString()}`}</span>
+	{/if}
 	<span class="currency">{currency}</span>
 </div>