+++ src/lib/components/Forms/InputWrapper.svelte
@@ -15,10 +15,11 @@
 	export let isValid: boolean;
 	export let tooltip: string | undefined = undefined;
 	export let statusIcons = true;
+	export let disabled = false;
 </script>
 
 {#if label}
-	<label for={name} class="row-2">
+	<label for={name} class="row-2" class:disabled>
 		{label}
 		{#if tooltip}
 			<TooltipIcon {tooltip} width={0.75} />