+++ src/lib/components/Label/Label.svelte
@@ -1,14 +1,16 @@
 <script type="ts">
 	import Icon from '@iconify/svelte';
 
-	export let color: 'primary' | 'secondary' | 'tertiary' | 'neutral' | 'transparent' = 'primary';
-	export let size: 'x-small' | 'small' | 'medium' | 'large' = 'medium';
+	export let color: 'primary' | 'secondary' | 'tertiary' | 'neutral' | 'transparent' | 'alert' =
+		'primary';
+	export let size: 'xx-small' | 'x-small' | 'small' | 'medium' | 'large' = 'medium';
 	export let state: 'on' | 'off' = 'on';
 	export let iconLeft: string | undefined = undefined;
 	export let iconRight: string | undefined = undefined;
+	export let hasBorder = true;
 </script>
 
-<div class={`center ${color} ${size} ${state}`}>
+<div class={`center ${color} ${size} ${state}`} class:no-border={!hasBorder}>
 	{#if iconLeft}
 		<Icon icon={iconLeft} />
 	{/if}
@@ -30,7 +32,7 @@
 	.primary {
 		background-color: var(--clr-primary-badge);
 		color: var(--clr-primary-main);
-		border: 1px var(--clr-primary-main) solid;
+		border: 0.5px var(--clr-primary-main) solid;
 	}
 	.primary.off {
 		color: var(--clr-primary-off);
@@ -39,7 +41,7 @@
 	.secondary {
 		background-color: var(--clr-secondary-badge);
 		color: var(--clr-secondary-main);
-		border: 1px var(--clr-secondary-main) solid;
+		border: 0.5px var(--clr-secondary-main) solid;
 	}
 	.secondary.off {
 		color: var(--clr-secondary-off);
@@ -48,7 +50,7 @@
 	.tertiary {
 		background-color: var(--clr-tertiary-badge);
 		color: var(--clr-tertiary-main);
-		border: 1px var(--clr-tertiary-main) solid;
+		border: 0.5px var(--clr-tertiary-main) solid;
 	}
 	.tertiary.off {
 		color: var(--clr-tertiary-off);
@@ -57,7 +59,7 @@
 	.neutral {
 		background-color: var(--clr-neutral-badge);
 		color: var(--clr-text-main);
-		border: 1px var(--clr-border-primary) solid;
+		border: 0.5px var(--clr-border-primary) solid;
 	}
 	.neutral.off {
 		color: var(--clr-heading-off);
@@ -66,10 +68,24 @@
 	.transparent {
 		background-color: transparent;
 		color: var(--clr-text-main);
-		border: 1px var(--clr-border-primary) solid;
+		border: 0.5px var(--clr-border-primary) solid;
 	}
 	.transparent.off {
 		color: var(--clr-heading-off);
+	}
+
+	.alert {
+		background-color: var(--clr-alert-badge);
+		color: var(--clr-alert-main);
+		border: 0.5px var(--clr-alert-main) solid;
+	}
+	.alert.off {
+		color: var(--clr-alert-off);
+	}
+
+	.xx-small {
+		font-size: calc(var(--font-size-0) - 1px);
+		padding: 0 0.7em;
 	}
 
 	.x-small {
@@ -86,5 +102,9 @@
 
 	.large {
 		font-size: var(--font-size-3);
+	}
+
+	.no-border {
+		border: none;
 	}
 </style>